split models in own files
