+++ app/src/app/shared/models/artsearch.interface.ts
@@ -1,0 +1,9 @@
+
+export interface ArtSearch {
+    motifs?: string[];
+    artists?: string[];
+    movements?: string[];
+    genres?: string[];
+    materials?: string[];
+    locations?: string[];
+}
