+++ app/src/app/shared/models/genre.interface.ts
@@ -1,0 +1,5 @@
+import { Entity } from './entity.interface';
+
+export interface Genre extends Entity {
+    type: 'genre';
+}
