+++ app/src/app/shared/models/location.interface.ts
@@ -1,0 +1,10 @@
+import { Entity } from './entity.interface';
+
+export interface Location extends Entity {
+    country?: string;
+    website?: string;
+    part_of: Partial<Location>[];
+    lat?: string;
+    lon?: string;
+    type: 'location';
+}
