+++ app/src/app/shared/models/material.interface.ts
@@ -1,0 +1,5 @@
+import { Entity } from './entity.interface';
+
+export interface Material extends Entity {
+    type: 'material';
+}
