+++ app/src/app/shared/models/motif.interface.ts
@@ -1,0 +1,5 @@
+import { Entity } from './entity.interface';
+
+export interface Motif extends Entity {
+    type: 'motif';
+}
