+++ app/src/app/shared/models/movement.interface.ts
@@ -1,0 +1,6 @@
+import { Entity } from './entity.interface';
+
+export interface Movement extends Entity {
+    influenced_by: Partial<Entity>[];
+    type: 'movement';
+}
