+++ app/src/app/shared/models/tagitem.interface.ts
@@ -1,0 +1,6 @@
+
+export interface TagItem {
+    label: string;
+    type?: string;
+    id?: string;
+}
